reorganised common files
