+++ src/generators/step_tensor.py
@@ -5,10 +5,12 @@ from pathlib import Path
 import network_diffusion as nd
 from tqdm import tqdm
 
+from src import os_utils, sim_utils
 from src.icm.torch_model import TorchMICModel, TorchMICSimulator
-from src.generators import commons
-from src.generators.utils import mean_simulation_results, SimulationResult
-from src.utils import export_dataclasses
+from src.generators.utils import(
+        SimulationResult,
+        mean_simulation_results,
+)
 
 
 def experiment_step(
@@ -36,7 +38,7 @@ def experiment_step(
 
             # update progress_bar
             p_bar.set_description_str(
-                commons.get_case_name_rich(
+                sim_utils.get_case_name_rich(
                     protocol=protocol,
                     p=p,
                     net_name=net_name,
@@ -76,5 +78,5 @@ def experiment_step(
             marginal_efficiencies.extend(repeated_results)
 
     # save efficiences obtained for this case
-    investigated_case_file_path = out_dir / f"{commons.get_case_name_base(protocol, p, net_name)}.csv"
-    export_dataclasses(marginal_efficiencies, investigated_case_file_path)
+    out_path = out_dir / f"{sim_utils.get_case_name_base(protocol, p, net_name)}.csv"
+    os_utils.export_dataclasses(marginal_efficiencies, out_path)
